# src/docs/json/swagger.json
{"swagger":"2.0","info":{"description":"BIMFACE 服务端API","version":"1.0","title":"Spring Boot中使用Swagger2构建RESTful APIs","termsOfService":"http://bimface.com/","contact":{"name":"zyj"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"hello-controller","description":"hello Controller"},{"name":"task-controller","description":"task Controller"},{"name":"test-controller","description":"测试相关api"}],"paths":{"/hello":{"get":{"tags":["hello-controller"],"summary":"文件","description":"到底该怎么解决啊！！！","operationId":"helloUsingGET_7","produces":["*/*","application/json"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/task/":{"get":{"tags":["task-controller"],"summary":"findAllTask","operationId":"findAllTaskUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/test/file":{"get":{"tags":["test-controller"],"summary":"文件类型","description":"获取支持文件类型","operationId":"fileSupportUsingGET","produces":["*/*","application/json"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/user对象"}},"400":{"description":"Invalid ID supplied","headers":{"支持类型":{"type":"string"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Pet not found","headers":{"支持类型":{"type":"string"}}}},"deprecated":false,"test1":"value1","test2":"value2"}},"/test/fileMetadata/{fileId}":{"get":{"tags":["test-controller"],"summary":"文件","description":"获取文件信息","operationId":"fileMetadataUsingGET","produces":["*/*","application/json"],"parameters":[{"name":"TOKEN","in":"header","description":"Authorization token","required":true,"type":"string"},{"name":"fileId","in":"path","description":"文件的id","required":true,"type":"string"}],"responses":{"200":{"description":"success","headers":{"X-Rack-Cache":{"type":"string","description":"Explains whether or not a cache was used"}},"schema":{"$ref":"#/definitions/FileBean"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Pet not found"}},"deprecated":false}},"/test/getCaV1":{"get":{"tags":["test-controller"],"summary":"getCaV1","operationId":"getCaV1UsingGET","produces":["*/*"],"parameters":[{"name":"fileId","in":"query","description":"fileId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/CategoryBean"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/test/getCaV2":{"get":{"tags":["test-controller"],"summary":"getCaV2","operationId":"getCaV2UsingGET","produces":["*/*"],"parameters":[{"name":"fileId","in":"query","description":"fileId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/TreeNode"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/test/getDownloadUrl":{"get":{"tags":["test-controller"],"summary":"getDownloadUrl","description":"An item can be of different type:\n\n type | definition\n -----|-----------\n Vinyl| aa\n VHS  | bbAn item can be of different type:\n\n type | definition\n -----|-----------\n Vinyl| aa\n VHS  | bbAn item can be of different type:\n\n type | definition\n -----|-----------\n Vinyl| aa\n VHS  | bb","operationId":"getDownloadUrlUsingGET","produces":["*/*","application/json"],"parameters":[{"name":"fileId","in":"query","description":"fileId","required":true,"type":"integer","format":"int64"},{"name":"fileName","in":"query","description":"fileName","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/test/getElements":{"get":{"tags":["test-controller"],"summary":"getElements","operationId":"getElementsUsingGET","produces":["*/*"],"parameters":[{"name":"categoryId","in":"query","description":"categoryId","required":false,"type":"string"},{"name":"family","in":"query","description":"family","required":false,"type":"string"},{"name":"familyType","in":"query","description":"familyType","required":false,"type":"string"},{"name":"fileId","in":"query","description":"fileId","required":false,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"string"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/test/getElements1":{"post":{"tags":["test-controller"],"summary":"getUs","operationId":"getUsUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"user","description":"user","required":true,"schema":{"$ref":"#/definitions/user对象"}}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"string"}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/test/getFloor":{"get":{"tags":["test-controller"],"summary":"getFloor","operationId":"getFloorUsingGET","produces":["*/*"],"parameters":[{"name":"integrateId","in":"query","description":"integrateId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/FloorTree"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/test/getPolicy":{"get":{"tags":["test-controller"],"summary":"getPolicy","operationId":"getPolicyUsingGET","produces":["*/*"],"parameters":[{"name":"name","in":"query","description":"name","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UploadPolicyBean"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/test/getShare":{"get":{"tags":["test-controller"],"summary":"createShare","operationId":"createShareUsingGET","produces":["*/*"],"parameters":[{"name":"fileId","in":"query","description":"fileId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ShareLinkBean"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/test/getSpecialty":{"get":{"tags":["test-controller"],"summary":"getSpecialty","operationId":"getSpecialtyUsingGET","produces":["*/*"],"parameters":[{"name":"integrateId","in":"query","description":"integrateId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/SpecialtyTree"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/test/getViewToken":{"get":{"tags":["test-controller"],"summary":"getViewToken","operationId":"getViewTokenUsingGET","produces":["*/*"],"parameters":[{"name":"fileId","in":"query","description":"fileId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}}},"definitions":{"CategoryBean":{"type":"object","properties":{"categoryId":{"type":"string"},"categoryName":{"type":"string"},"families":{"type":"array","items":{"$ref":"#/definitions/Family"}}},"title":"CategoryBean"},"Family":{"type":"object","properties":{"family":{"type":"string"},"familyTypes":{"type":"array","items":{"type":"string"}}},"title":"Family"},"FileBean":{"type":"object","properties":{"createTime":{"type":"string"},"etag":{"type":"string"},"fileId":{"type":"integer","format":"int64"},"length":{"type":"integer","format":"int64"},"name":{"type":"string"},"status":{"type":"string"},"suffix":{"type":"string"}},"title":"FileBean"},"Floor":{"type":"object","properties":{"categories":{"type":"array","items":{"$ref":"#/definitions/CategoryBean"}},"floor":{"type":"string"},"specialties":{"type":"array","items":{"$ref":"#/definitions/Specialty"}}},"title":"Floor"},"FloorTree":{"type":"object","properties":{"tree":{"type":"array","items":{"$ref":"#/definitions/Floor"}},"treeType":{"type":"integer","format":"int32"}},"title":"FloorTree"},"ShareLinkBean":{"type":"object","properties":{"expireTime":{"type":"string"},"url":{"type":"string"}},"title":"ShareLinkBean"},"Specialty":{"type":"object","properties":{"categories":{"type":"array","items":{"$ref":"#/definitions/CategoryBean"}},"floors":{"type":"array","items":{"$ref":"#/definitions/Floor"}},"specialty":{"type":"string"}},"title":"Specialty"},"SpecialtyTree":{"type":"object","properties":{"tree":{"type":"array","items":{"$ref":"#/definitions/Specialty"}},"treeType":{"type":"integer","format":"int32"}},"title":"SpecialtyTree"},"SupportFileBean":{"type":"object","properties":{"length":{"type":"integer","format":"int64"},"types":{"type":"array","items":{"type":"string"}}},"title":"SupportFileBean"},"TreeNode":{"type":"object","properties":{"elementCount":{"type":"integer","format":"int32"},"id":{"type":"string"},"items":{"type":"array","items":{"$ref":"#/definitions/TreeNode"}},"name":{"type":"string"},"type":{"type":"string"}},"title":"TreeNode"},"UploadPolicyBean":{"type":"object","properties":{"accessId":{"type":"string"},"callbackBody":{"type":"string"},"expire":{"type":"integer","format":"int64"},"host":{"type":"string"},"objectKey":{"type":"string"},"policy":{"type":"string"},"signature":{"type":"string"}},"title":"UploadPolicyBean"},"user对象":{"type":"object","properties":{"age":{"type":"integer","format":"int32","example":21,"description":"年龄"},"id":{"type":"integer","format":"int64","example":21,"description":"id"},"name":{"type":"string","example":"xingguo","description":"用户名"}},"title":"user对象","description":"用户对象user"}}}